Refactor to trueval agent single and base
